Minor refactoring

+++ Source/Aonir/Core/Window/WindowSettings.hpp
@@ -1,22 +1,23 @@
 #pragma once
 
 #include <compare>
+#include <cstdint>
 #include <string>
 
 namespace Aonir
 {
     struct WindowPosition
     {
-        std::size_t x;
-        std::size_t y;
+        std::uint32_t x;
+        std::uint32_t y;
 
         auto operator<=>(const WindowPosition &) const = default;
     };
 
     struct WindowSize
     {
-        std::size_t width;
-        std::size_t height;
+        std::uint32_t width;
+        std::uint32_t height;
 
         auto operator<=>(const WindowSize &) const = default;
     };
